rename package names
