+++ pkg/test/useraccount/assertion.go
@@ -1,4 +1,4 @@
-package uatest
+package useraccount
 
 import (
 	"context"
